requirements.txt
Django==1.9.2
dj-database-url==0.4.0
django-admin-tools==0.5.2
django-appconf==1.0.1
django-extra-views==0.6.4
django-google-charts==0.1.1
django-haystack==2.4.1
django-imagekit==3.3
django-qsstats-magic==0.7.2
django-sphinx==2.2.4
django-tables2==1.0.4
django-treebeard==3.0
django-widget-tweaks==1.4.1
MySQL-python==1.2.5
mysqlclient==1.3.7
Pillow==2.8.2
PyMySQL==0.6.7
python-dateutil==1.5